piccache.php: Do not write empty files when image download fails
When an image download fails, an empty file is currently written,
which is meaningless.
Also, the download failure may be temporary (for example, the service
is temporarily unavailable due to network reasons). The existence
of an empty file means that the user will never be able to try to
retrieve the file again.

When the download fails, do not write an empty file so that users
can try to retrieve it again when they access it.
Additionally, a timeout for downloading images has been added to
avoid long blocking times.